// src/interfaces/wms/supplier_wms.interface.ts
export interface ISupplier {
  company_code?: string;
  prin_code: string;
  supp_code: string;
  curr_code?: string;
  country_code?: string;
  supp_name?: string;
  supp_addr1?: string;
  supp_addr2?: string;
  supp_addr3?: string;
  supp_addr4?: string;
  supp_city?: string;
  supp_contact1?: string;
  supp_telno1?: string;
  supp_faxno1?: string;
  supp_email1?: string;
  supp_contact2?: string;
  supp_telno2?: string;
  supp_faxno2?: string;
  supp_email2?: string;
  supp_contact3?: string;
  supp_telno3?: string;
  supp_faxno3?: string;
  supp_ref1?: string;
  supp_ref2?: string;
  supp_ref3?: string;
  service_date?: Date;
  supp_acref?: string;
  supp_credit?: number;
  supp_stat?: string;
  supp_imp_code?: string;
  supp_lic_no?: string;
  supp_lic_type?: string;
  price_check?: string;
  supp_email3?: string;
  payment_terms?: number;
  importer_code?: string;
  updated_at?: Date;
  updated_by?: string;
  created_by?: string;
  created_at?: Date;
}